Module buildconfig renamed to configfile
