+++ lib/build.js
@@ -2,7 +2,7 @@ var async = require('async');
 var fse = require('fs-extra');
 var log = require('npmlog');
 
-var buildconfig = require('./buildconfig');
+var configfile = require('./configfile');
 var closure = require('./index');
 var config = require('./config');
 
@@ -22,7 +22,7 @@ log.level = config.get('log_level');
  */
 function _readConfig(configFile, callback) {
   log.info('closure-util', 'Reading build config');
-  buildconfig.readConfig(configFile, callback);
+  configfile.readConfig(configFile, callback);
 }
