+++ lib/configfile.js
