+++ lib/serve.js
@@ -1,8 +1,8 @@
 var async = require('async');
 var log = require('npmlog');
 
-var buildconfig = require('./buildconfig');
 var config = require('./config');
+var configfile = require('./configfile');
 var manager = require('./manager');
 var server = require('./server');
 
@@ -22,7 +22,7 @@ log.level = config.get('log_level');
  */
 function _readConfig(configFile, callback) {
   log.info('closure-util', 'Reading build config');
-  buildconfig.readConfig(configFile, callback);
+  configfile.readConfig(configFile, callback);
 }
